third_party/blink/web_tests/external/wpt/editing/run/delete_4001-5000-expected.txt

This is a testharness.js-based test.
Found 18 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["delete",""]] "<p>foo<p style=text-decoration:line-through>[]bar" queryCommandState("stylewithcss") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["delete",""]] "<p>foo<p style=text-decoration:line-through>[]bar" queryCommandValue("defaultparagraphseparator") before
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["delete",""]] "<p>foo<p style=text-decoration:line-through>[]bar" queryCommandValue("defaultparagraphseparator") after
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "<p><u>foo</u><p>[]bar" queryCommandValue("defaultparagraphseparator") before
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["stylewithcss","true"],["delete",""]] "<p style=color:blue>foo</p>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"color:rgb(0, 0, 255)\\">foo</span>bar</p>" but got "<p style=\\"color:rgb(0, 0, 255)\\">foobar</p>"
[FAIL] [["stylewithcss","true"],["delete",""]] "<p style=color:blue>foo</p>[]bar" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["delete",""]] "<p style=color:blue>foo</p>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font color=\\"#0000ff\\">foo</font>bar</p>" but got "<p style=\\"color:rgb(0, 0, 255)\\">foobar</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["delete",""]] "foo<p style=color:brown>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font color=\\"#a52a2a\\">bar</font>" but got "foo<span style=\\"color:rgb(165, 42, 42)\\">bar</span>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["delete",""]] "foo<p style=color:brown>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font color=\\"#a52a2a\\">bar</font>" but got "foo<span style=\\"color:rgb(165, 42, 42)\\">bar</span>"
[FAIL] [["stylewithcss","true"],["delete",""]] "<div style=color:blue><p style=color:green>foo</div>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><p><span style=\\"color:rgb(0, 128, 0)\\">foo</span>bar</p></div>" but got "<div style=\\"color:rgb(0, 0, 255)\\"><p style=\\"color:rgb(0, 128, 0)\\">foobar</p></div>"
[FAIL] [["stylewithcss","false"],["delete",""]] "<div style=color:blue><p style=color:green>foo</div>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><p><font color=\\"#008000\\">foo</font>bar</p></div>" but got "<div style=\\"color:rgb(0, 0, 255)\\"><p style=\\"color:rgb(0, 128, 0)\\">foobar</p></div>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["delete",""]] "<div style=color:blue><p style=color:green>foo</div><p style=color:brown>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"color:rgb(0, 0, 255)\\"><p style=\\"color:rgb(0, 128, 0)\\">foo<font color=\\"#a52a2a\\">bar</font></p></div>" but got "<div style=\\"color:rgb(0, 0, 255)\\"><p style=\\"color:rgb(0, 128, 0)\\">foo<span style=\\"color:rgb(165, 42, 42)\\">bar</span></p></div>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["delete",""]] "<div style=color:blue><p style=color:green>foo</div><p style=color:brown>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"color:rgb(0, 0, 255)\\"><p style=\\"color:rgb(0, 128, 0)\\">foo<font color=\\"#a52a2a\\">bar</font></p></div>" but got "<div style=\\"color:rgb(0, 0, 255)\\"><p style=\\"color:rgb(0, 128, 0)\\">foo<span style=\\"color:rgb(165, 42, 42)\\">bar</span></p></div>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["delete",""]] "<p style=color:blue>foo<div style=color:brown><p style=color:green>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"color:rgb(0, 0, 255)\\">foo<font color=\\"#008000\\">bar</font></p>" but got "<p style=\\"color:rgb(0, 0, 255)\\">foo<span style=\\"color:rgb(0, 128, 0)\\">bar</span></p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["delete",""]] "<p style=color:blue>foo<div style=color:brown><p style=color:green>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"color:rgb(0, 0, 255)\\">foo<font color=\\"#008000\\">bar</font></p>" but got "<p style=\\"color:rgb(0, 0, 255)\\">foo<span style=\\"color:rgb(0, 128, 0)\\">bar</span></p>"
[FAIL] [["delete",""]] "<p>foo</p><p>{bar</p>}<p>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p><br></p><p>baz</p>" but got "<p>foo</p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "<p>foo</p>{<p>bar}</p><p>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><br><p>baz</p>" but got "<p>foo</p><p><br></p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "<p>foo</p>{<p>bar}</p><p>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><br><p>baz</p>" but got "<p>foo</p><p><br></p><p>baz</p>"
Harness: the test ran to completion.
